[IMP] typ: Validating users can't edit Partner from SO with Lines
Once the sale order is created, we want to avoid users to edit the
partner field to avoid bad price assignation to lines or unexpected
changes on the order.

The only way to be able to change this information on the SOs is
the order has no lines on it.