[FIX] typ: Ensure invoice is not being signed when sending it by e-mail
When sending an invoice by e-mail, it may be in the process of being
signed, e.g. when an invoice generated from the PoS is sent
automatically.

To solve the above, the invoice is now sent after pending processes are
finished (AKA committed), similar as natively done by the mexican
locatlization.