[REM] typ: Rendering of POS dependencies
There were missing dependencies and Non loaded modules when opening the
POS view, because now those files were migrated to OWL and at least
"Screens" have different files. They are being removed from here until
there is a migration to them to OWL, and that should be in the mexico
repo not here.