[REM] typ: Extra field serial in stock move lines
In 11.0, there was a copy of the native field for lots (`lot_id`) called
Serial (`serial_id`). The purpose of that field was to ease scanning
serials when using the barcode and to avoid some issues whenenterin
 already-present serials.

Since we're sticking as much as possible to Odoo standard, specially
barcode-related features where all custom features were removed (at least
for now), the aforementioned field is removed.