[REM] typ: feature to warng about partner's fiscal position
This removes a feature that was intended to warng the user when a
customer is selected on a sale order, whose fiscal position differs from
the one set in the sales team.

As accorded with the customer, the above feature is not useful anymore,
as the fiscal position will always be taken from the sales team, no
matters the customer.