[FIX] typ: Rules by sales team shouldn't apply for portal users
There are several record rules to limit access to some documents based on
the current user's sales team:
- Sale orders
- Purchase orders
- Journal entries
- Journal items

However, such rules should only apply to internal users, as those are
the only ones intended to belong to sales teams (portal users are not
sellers). In addition, native rules already ensure portal users are
able to see only their own documents.

To solve the above, a new condition is added to ensure the restriction
only applies for internal users (i.e. users that are not shared).

Note:
Rules were not linked to the portal group because we need them to be
global, as it's the only way to ensure the restriction is applied even
if there are other rules that grant additional access (e.g. the native
ones).