[ADD] typ: Migration script to move supplierinfo -> customerinfo
Since v12.0, supplierinfos for customers are handled on a separate
model. Such change is handled in the migration script in 12.0, on its
module (`product_supplierinfo_for_customer`) [1]. However, since we're
migrating from 11.0 to 14.0, such script is not being executed.

The above is fixed by copying the migration script into the main app.

[1] https://github.com/Vauxoo/product-attribute/blob/12.0/product_supplierinfo_for_customer/migrations/12.0.1.0.0/pre-migration.py