[FIX] typ: Invoice Report Font Size (Portal)
Adding the correct size in some of the fonts in the invoice report,
because in Portal it was missing the Unit so the `font-size` element was
not really being set.
Also, in the sections "Issuer", "Customer" and "Issued at" padding was
added to the sides so the text wouldn't be so close to the borders, and
we remove some line breaks that were unnecessary.