[MIG] typ: feature to create purchase order from sale order
In v11.0, there's a feature called "Special sale order", which allows
to create a purchase order from a sale order, allowing to define a
specific vendor.

Odoo natively supports creating a purchase order by using routes, but
it doesn't allow to select a vendor.

This commit migrates the aforementioned feature.