[REM] typ: feature to edit credit limit when there's one warehouse
Credit limit field is warehouse-dependent and its behavior changed
depending on how many warehouse configurations were present:
- When there wasn't any:
  * field was editable like a standard field
- When there was exactly one:
  * Field was editable
  * When edited, its value was written to the warehouse config (through
    an inverse)
  * When the warehouse config was edited, field was also updated
- When there were multiple:
  * field was readonly
  * It was computed by summing credit in all warehouses

However, we were having issues when there was one config and the config
was updated, because the credit was not being updated immediately. ANd
fixing that would make harder to have both consistent with the same value,
because each one would want to update the other.

To solve the above:
- Field is editable only for one config, and computed otherwise (removed inverse)
- Fixed computing so it updates credit limit immediately
- Removed field to count configs as it's not required anymore for the
  new logic, it was used only to make credit limit readonly
- When creating first config, default credit is taken from partner