[REF] typ: don't consider sales when computing available credit
Even though we want credit warnings to be shown in sale orders, we
don't want sales amount to be considered when computing available credit
amount. That's because there are several special cases where sale
orders are confirmed but delayed, e.g. special sale orders (the ones
that create purchase orders).