[FIX] typ: Vendor for purchase not considered in sale order lines
When creating a special sale order (i.e. a sale order that triggers a
purchase order), the specified vendor was only considered when the
purchase was triggered by a reordering rule. That's not a valid case
for us anyway, as it doesn't consider specified rule.

This commit performs the following changes:
- Fix vendor calculation so the one in the sale order line is considered
- Ensure inventory scheduler is not automatically run when reserving
  products

The vendor assignation works as follows:
- A new field is added to stock moves to store the vendor for purchases
- The field is  filled from sale orders
- It's propagated across chained moves (move -> procurement -> origin move)
- It's used when creating purchase orders, when a buy rule is run

Even though not strictly required, field is initialized on old records,
for consistency and auditing purposes.