[FIX] typ: access error when canceling picking coming from an SO
When canceling pickings, we check related sale order to:
- Cancel order if all other pickings are already cancelled
- Check if it's a special sale order, to prevent cancellation in certain
  cases

However, that could lead to an access error if the current user has
access to cancel the current picking but not to read related sale orders.

To solve the above, related sale orders are read as superuser, and may
only be cancelled if current user has access to cancel the picking.