[FIX] typ: access error when cancelling last chained picking
When chained pickings are created and they involve several warehouses
(e.g. through a manual transfer), if an user with access to only
the last picking tries to cancel it, an access error is raised. That's
because we have a feature to, when a chained picking is cancelled, a
message is posted to source pickings, containing a list of the ones that
were cancelled.

This change solves the above by ensuring chained pickings are looked up
and posted as superuser.

In addition, the posted message is improved to be an actual HTML list.