[FIX] typ: wrong invoice amounts coming from migration
Some invoice lines were not fully migrated to journal items because,
prior to migration, values of journal items didn't match the ones on
invoice lines. So Odoo just created new journal items of type "Section"
that are ignored.

A migration script is implemented to fix invoice amount on those cases.